% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepaf.R
\name{decodePolyline}
\alias{decodePolyline}
\title{Decode a Google Polyline to a Data Frame}
\usage{
decodePolyline(enc_polyline, factor = 5)
}
\arguments{
\item{enc_polyline}{a Google polyline.}

\item{factor}{number of decimal digits to be used.}
}
\value{
A data frame of latitudes and longitudes is returned.
}
\description{
Decode a Google polyline to a data frame of coordinates.
}
\examples{
coords <- decodePolyline(enc_polyline = "_p~iF~ps|U_ulLnnqC_mqNvxq`@")
coords
}

