% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepaf.R
\name{encodePolyline}
\alias{encodePolyline}
\title{Encode Coordinates to Google Polylines}
\usage{
encodePolyline(df_coords, factor = 5)
}
\arguments{
\item{df_coords}{a data frame of coordinates with two columns: latitudes and
longitudes. Coordinates must be in decimal degrees (WGS84).}

\item{factor}{number of decimal digits to be used.}
}
\value{
An encoded polyline is returned.
}
\description{
Encode a data.frame of coordinates to a Google polyline.
}
\examples{
coords <- data.frame(lat = c(38.5, 40.7, 43.252),
                     lon = c(-120.2, -120.95, -126.453))
encpoly <- encodePolyline(coords)
encpoly
}

