% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gepaf}
\alias{gepaf}
\alias{gepaf-package}
\title{Google Encoded Polyline Algorithm Format}
\description{
Encode and decode the Google Encoded Polyline Algorithm Format.\cr
See the documentation here:\cr
\url{https://developers.google.com/maps/documentation/utilities/polylinealgorithm}.
}
\note{
Mostly a translation of \url{http://github.com/mthh/polyline_ggl/} (itself a
modest translation (i.g no GeoJSON wrapper, etc.) of other well known available
implementations such as \href{https://www.npmjs.com/package/polyline}{Node.js Mapbox} one).
}
\seealso{
\link{encodePolyline}, \link{decodePolyline}.
}

